+++ exporter/opentelemetry-exporter-otlp/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add Gzip compression for exporter
+  ([#1141](https://github.com/open-telemetry/opentelemetry-python/pull/1141))
 ## Version 0.15b0
 
 Released 2020-11-02